pyproject.toml
[project]
name = "pyfound"
version = "0.1.3"
description = "Found! is an AI enabled tool for fast semantic search over your personal documents."
readme = "README.md"
authors = [
    { name = "Henning Claßen", email = "code@clssn.de" }
]
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "pandas>=2.3.1",
    "platformdirs>=4.3.8",
    "sentence-transformers>=5.0.0",
    "typer>=0.16.0",
]

[project.scripts]
found = "found:cli"
pyfound = "found:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/found"]

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
